Add JWT Authentication, move logic from models.py feat: Add JWT Bearer Authentication   - /auth/login             to get bearer token   - /auth/login/who-am-i    to return user id
Fix bug where user can send sign up request for existing users

* Update sign up so user can send multiple requests
* Update error messages to reduce chances for user enumeration
* Update sign up email message and config.py so it links to frontend url
* Create a exceptions.py file for some app specific exceptions

+++ backend/ezConnect/auth/signup.py
@@ -1,8 +1,9 @@
 # from flask import Config
 from datetime import datetime, timedelta
-from ezConnect.config import DOMAINS, HOSTNAME
+from ezConnect.config import DOMAINS, FRONTEND_HOSTNAME
 from ezConnect.utils.emailer import send_email
 from ezConnect.models import User, SignUpRequest, db
+from ezConnect.auth.login import _hash_password, _get_user
 
 import traceback
 
@@ -11,7 +12,7 @@ def request(body):
     req = {
         "creation_time" : datetime.now(),
         "email" : body['email'],
-        "namne" : "john doe"
+        "name" : "john doe"
     }
     # first validate email
     email: str = body['email']
@@ -19,13 +20,21 @@ def request(body):
         return {"error": "Not an email"}, 400
     if email.split('@')[-1] not in DOMAINS:
         return {"error": "Invalid domain"}, 401
-    
+    # Check duplicate, do not send request if the email is already being used
+    user = _get_user(email)
+    success_msg = f"Verification email sent to {req['email']} if account can be created"
+    if user is not None:
+        # Do not disclose if account / email exist to the requester
+        # See: Account Enumeration
+        print(f"Request made for existing account {email}")
+        return {"message" : success_msg}, 200
+
     try: 
         sign_up_request = SignUpRequest(datetime.utcnow(), email)
         db.session.add(sign_up_request)
         db.session.commit()
         sign_up_token = sign_up_request.id
-        message= f"Please go to https://{HOSTNAME}/signup/verify?signUpRequestToken={sign_up_token}" + \
+        message= f"Please go to https://{FRONTEND_HOSTNAME}/signup/verify?signUpRequestToken={sign_up_token}" + \
             f"&email={email} to verify your account and continue with account creation"
         print(f"{message}")
         send_email(
@@ -38,7 +47,7 @@ def request(body):
         traceback.print_exc()
         return {"error": f"str(e)"}, 500
 
-    return {"message" : f"Verification email sent to {req['email']}"}, 200
+    return {"message" : success_msg}, 200
 
 def verify(emailToken, emailQueryString, body):
     sign_up_req: SignUpRequest = SignUpRequest.query.get(emailToken)
@@ -46,20 +55,28 @@ def verify(emailToken, emailQueryString, body):
         return {"error": "Invalid email verification token"}, 404
     if emailQueryString != sign_up_req.email:
         return {"error": "Invalid email for this verification token"}, 401
+    password: str = body['password']
+    if len(password) < 12:
+        return {"error": "Password too short"}, 422
     try:
         # Past 24 hours, delete
         if datetime.utcnow() - sign_up_req.creation_time >  timedelta(days=1):
             # Expired verification email token
             db.session.delete(sign_up_req)
             db.session.commit()
-            return {"message": "Email verification token expired"}, 401
+            return {"message": "Email verification token expired, please make a new request"}, 401
         email_addr = emailQueryString
+        salt, password_hash = _hash_password(password)
+        # TODO: Update the course to reflect new structure after data normalisation
         new_user = User(name=body['name'], 
                         email=email_addr,
-                        password=body['password'],
+                        salt = salt,
+                        password=password_hash,
                         year=body['year'],
                         course=body['course'])
-        db.session.delete(sign_up_req)
+        # db.session.delete(sign_up_req)
+        # Delete other requests that might have been made too(
+        SignUpRequest.query.filter_by(email=email_addr).delete()
         db.session.add(new_user)
         db.session.commit()
         return {"message": f"User {new_user.name}" + \
